Migrated currency values to BigDecimal

+++ feature-currency-feed/src/main/kotlin/com/bradley/wilson/currency/feed/CurrencyItem.kt
@@ -1,12 +1,14 @@
 package com.bradley.wilson.currency.feed
 
+import java.math.BigDecimal
+
 data class Currency(
     val country: String,
-    val rate: Double
+    val rate: BigDecimal
 )
 
 data class CurrencyItem(
     val country: String,
-    val rate: Double,
+    val rate: BigDecimal,
     val isBateRate: Boolean = false
 )
